packaging: making all programmatically callable
That's a big commit:
 - Creating a subdirectory called glouton/ where all the code resides
 - Creating setup.py and setup.cfg
