--- commands/module/moduleCommand.py
@@ -1,9 +1,0 @@
-from infrastructure.satnogClient import SatnogClient
-import os
-
-class ModuleCommand:
-    def __init__(self, params):
-        self.__params = params
-
-    def process(self):
-        self.__params.module.runAfterDownload(self.__params.file_name, self.__params.full_path, self.__params.observation)
