+++ glouton/shared/fileHelper.py
