+++ setup.cfg
@@ -1,0 +1,6 @@
+[bdist_wheel]
+universal=1
+
+[metadata]
+description-file = README.md
+
